requirements.txt
gradio
google-genai
PyMuPDF
Pillow
python-dotenv
